backend/requirements.txt
numpy==1.26.2
matplotlib>=3.8,<4.0
pandas >= 2.2, <3.0
scikit-learn==1.3.2
joblib==1.3.2
fastapi==0.108.0
uvicorn==0.25.0
# pydantic-settings==2.1.0
pydantic-settings>=0.1
# pydantic-settings>=0.1
# psycopg2-binary==2.9.3
psycopg2-binary>=2.9.3
# pydantic>=1.8,<2
PyPDF2==1.26.0
python-jose==3.3.0
apitally[fastapi]==0.3.3
starlette==0.32.0
alembic==1.13.1
pyodbc==5.1.0 
# requests >=2.31,<3
python-multipart==0.0.5
SQLAlchemy==1.4.0
# psycopg2 ==2.8.6
ultralytics>=8.0.232
yolov5==6.1
openpyxl==3.1.2
Pillow==10.3.0
unstructured_client==0.22.0
unstructured==0.5.1
python-utils==3.8.2
python-dotenv==1.0.1
pytesseract==0.3.10
pdf2image==1.17.0
camelot-py==0.11.0
PyMuPDF==1.24.2
s3fs==2024.3.1
pyarrow==16.0.0
pdfplumber==0.11.0
ghostscript==0.7
surya-ocr==0.4.0
huggingface-hub==0.23.0
pytest-httpx==0.30.0